Move `OperationContext` to `linera-base`
Replace the duplicated types in `linera-execution` and `linera-sdk`, and
add a missing field to the WIT type.
